+++ Classes/BTLocationTask.h
@@ -6,8 +6,8 @@
 #import "BTInterpolationTask.h"
 
 @interface BTLocationTask : BTInterpolationTask
-- (id)initOverTime:(double)seconds toX:(float)x toY:(float)y;
-- (id)initOverTime:(double)seconds toX:(float)x toY:(float)y withInterpolator:(BTInterpolator)interp;
-- (id)initOverTime:(double)seconds toX:(float)x toY:(float)y onDisplay:(SPDisplayObject*)display;
-- (id)initOverTime:(double)seconds toX:(float)x toY:(float)y withInterpolator:(BTInterpolator)interp onDisplay:(SPDisplayObject*)display;
+- (id)initOverTime:(float)seconds toX:(float)x toY:(float)y;
+- (id)initOverTime:(float)seconds toX:(float)x toY:(float)y withInterpolator:(BTInterpolator)interp;
+- (id)initOverTime:(float)seconds toX:(float)x toY:(float)y onDisplay:(SPDisplayObject*)display;
+- (id)initOverTime:(float)seconds toX:(float)x toY:(float)y withInterpolator:(BTInterpolator)interp onDisplay:(SPDisplayObject*)display;
 @end
