+++ Classes/BTRotationTask.h
@@ -5,12 +5,12 @@
 #import "BTInterpolationTask.h"
 
 @interface BTRotationTask : BTInterpolationTask
-- (id)initOverTime:(double)seconds toRotation:(float)rads;
+- (id)initOverTime:(float)seconds toRotation:(float)rads;
 
-- (id)initOverTime:(double)seconds toRotation:(float)rads withInterpolator:(BTInterpolator)interp;
+- (id)initOverTime:(float)seconds toRotation:(float)rads withInterpolator:(BTInterpolator)interp;
 
-- (id)initOverTime:(double)seconds toRotation:(float)rads onDisplay:(SPDisplayObject*)display;
+- (id)initOverTime:(float)seconds toRotation:(float)rads onDisplay:(SPDisplayObject*)display;
 
-- (id)initOverTime:(double)seconds toRotation:(float)rads
+- (id)initOverTime:(float)seconds toRotation:(float)rads
   withInterpolator:(BTInterpolator)interp onDisplay:(SPDisplayObject*)display;
 @end
