+++ Classes/BTRotationTask.m
@@ -7,20 +7,20 @@
     float _startRotation;
     float _deltaRotation;
 }
-- (id)initOverTime:(double)seconds toRotation:(float)rads {
+- (id)initOverTime:(float)seconds toRotation:(float)rads {
     return [self initOverTime:seconds toRotation:rads withInterpolator:BTLinearInterpolator];
 }
 
-- (id)initOverTime:(double)seconds toRotation:(float)rads withInterpolator:(BTInterpolator)interp {
+- (id)initOverTime:(float)seconds toRotation:(float)rads withInterpolator:(BTInterpolator)interp {
     return [self initOverTime:seconds toRotation:rads withInterpolator:interp onDisplay:nil];
 }
 
-- (id)initOverTime:(double)seconds toRotation:(float)rads onDisplay:(SPDisplayObject*)display {
+- (id)initOverTime:(float)seconds toRotation:(float)rads onDisplay:(SPDisplayObject*)display {
     return [self initOverTime:seconds toRotation:rads withInterpolator:BTLinearInterpolator
                 onDisplay:display];
 }
 
-- (id)initOverTime:(double)seconds toRotation:(float)rads
+- (id)initOverTime:(float)seconds toRotation:(float)rads
   withInterpolator:(BTInterpolator)interp onDisplay:(SPDisplayObject*)display {
     if (!(self = [super initOverTime:seconds withInterpolator:interp onDisplay:display])) return nil;
     [self.attached connectUnit:^{
